Completed percent error fix
The completed percentage is coming more than 100%. The problem was - TestNG captures the excluded `@Test` methods list which are not really excluded by user, they are the configuration methods (i.e @BeforeTest etc..). So, now an extra check is applied from our code end to verify whether the `excluded` methods reported by TestNG is `@Test` annotated method or not.